Move TestFilters into test_utils
TestFilters is very useful for a variety of testing scenarios, so we're
moving it into our testing utility package so we can reuse it across
different modules.
